refactor : Changing the folder structure
