+++ lib/main.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import 'screen/home_page/home_page.dart';
+import 'portifolio/view/portfolio_page.dart';
+
 
 void main() {
   runApp(const MyApp());
@@ -16,7 +17,7 @@ class MyApp extends StatelessWidget {
       child: MaterialApp(
         debugShowCheckedModeBanner: false,
         title: 'Cripto',
-        home: HomePage(),
+        home: PortfolioPage(),
       ),
     );
   }
