+++ lib/movements/widgets/movements_body.dart
@@ -7,14 +7,12 @@ class MovementesBody extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: const [
-            Text("Ainda não implementado"),
-          ],
-        ),
+    return Center(
+      child: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: const [
+          Text("Ainda não implementado"),
+        ],
       ),
     );
   }
