+++ lib/portifolio/provider/utils.dart
@@ -1,0 +1,5 @@
+import 'package:hooks_riverpod/hooks_riverpod.dart';
+
+final visibleProvider = StateProvider(
+  (ref) => true,
+);
