+++ lib/portifolio/widgets/cripto_list_tile.dart
@@ -5,7 +5,8 @@ import 'package:decimal/intl.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../../../core/utils.dart';
+import '../../shared/utils/utils.dart';
+import '../provider/utils.dart';
 
 class CriptoListTile extends HookConsumerWidget {
   const CriptoListTile({
@@ -48,16 +49,16 @@ class CriptoListTile extends HookConsumerWidget {
                     sigmaY: 15,
                   ),
                   child: Text(
-                    ref.watch(numberFormatProvider).format(
-                          DecimalIntl(userMoneyBitcoin),
-                        ),
+                    numberFormat.format(
+                      DecimalIntl(userMoneyBitcoin),
+                    ),
                     style: const TextStyle(fontSize: 20),
                   ),
                 ),
                 child: Text(
-                  ref.watch(numberFormatProvider).format(
-                        DecimalIntl(userMoneyBitcoin),
-                      ),
+                  numberFormat.format(
+                    DecimalIntl(userMoneyBitcoin),
+                  ),
                   style: const TextStyle(fontSize: 20),
                 ),
               ),
