--- lib/screen/movements_page/movements_page.dart
@@ -1,12 +1,0 @@
-import 'package:flutter/material.dart';
-
-import 'widgets/movements_body.dart';
-
-class MovementsPage extends StatelessWidget {
-  const MovementsPage({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return const MovementesBody();
-  }
-}
