+++ lib/shared/utils/route_definitions.dart
@@ -1,0 +1,7 @@
+import '../../movements/view/movements_page.dart';
+import '../../portifolio/view/portfolio_page.dart';
+
+var routes = {
+  "/portifolio": (context) => const PortfolioPage(),
+  "/movements": (context) => const MovementsPage(),
+};
