+++ lib/shared/utils/utils.dart
@@ -1,0 +1,7 @@
+import 'package:intl/intl.dart';
+
+final numberFormat = NumberFormat.currency(
+  locale: "pt_BR",
+  decimalDigits: 2,
+  name: "R\$",
+);
